requirements.txt
asgiref==3.7.2
Django==4.2.7
django-extensions==3.2.3
django-modeltranslation==0.18.11
django-recaptcha==3.0.0
gunicorn==21.2.0
packaging==23.2
psycopg==3.1.13
psycopg-binary==3.1.13
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
